+++ packages/blocks/src/api/index.native.js
@@ -20,6 +20,8 @@ export {
 	getBlockTypes,
 	hasBlockSupport,
 	isReusableBlock,
+	setDefaultBlockName,
+	getDefaultBlockName,
 } from './registration';
 export { getPhrasingContentSchema } from './raw-handling';
 export { default as children } from './children';