+++ packages/data/src/test/registry.js
@@ -178,49 +178,40 @@ describe( 'createRegistry', () => {
 			expect( registry.select( 'butcher' ).getPrice( 'chicken' ) ).toBe( 2 );
 			expect( registry.select( 'butcher' ).getPrice( 'ribs' ) ).toBe( 6 );
 		} );
-	} );
 
-	// TODO: Refactor this into registerStore tests after this function is removed.
-	describe( 'registerReducer', () => {
 		it( 'Should append reducers to the state', () => {
 			const reducer1 = () => 'chicken';
 			const reducer2 = () => 'ribs';
 
-			const store = registry.registerReducer( 'red1', reducer1 );
+			const store = registry.registerStore( 'red1', { reducer: reducer1 } );
 			expect( store.getState() ).toEqual( 'chicken' );
 
-			const store2 = registry.registerReducer( 'red2', reducer2 );
+			const store2 = registry.registerStore( 'red2', { reducer: reducer2 } );
 			expect( store2.getState() ).toEqual( 'ribs' );
-
-			// This uses deprecated functions and will produce a warning.
-			expect( console ).toHaveWarned();
 		} );
-	} );
 
-	// TODO: Refactor this into registerStore tests after this function is removed.
-	describe( 'registerResolvers', () => {
 		it( 'should not do anything for selectors which do not have resolvers', () => {
-			registry.registerReducer( 'demo', ( state = 'OK' ) => state );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'OK' ) => state,
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {},
 			} );
-			registry.registerResolvers( 'demo', {} );
 
 			expect( registry.select( 'demo' ).getValue() ).toBe( 'OK' );
-
-			// This uses deprecated functions and will produce a warning.
-			expect( console ).toHaveWarned();
 		} );
 
 		it( 'should behave as a side effect for the given selector, with arguments', () => {
 			const resolver = jest.fn();
-
-			registry.registerReducer( 'demo', ( state = 'OK' ) => state );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: resolver,
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'OK' ) => state,
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: resolver,
+				},
 			} );
 
 			const value = registry.select( 'demo' ).getValue( 'arg1', 'arg2' );
@@ -234,13 +225,14 @@ describe( 'createRegistry', () => {
 
 		it( 'should support the object resolver definition', () => {
 			const resolver = jest.fn();
-
-			registry.registerReducer( 'demo', ( state = 'OK' ) => state );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: { fulfill: resolver },
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'OK' ) => state,
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: { fulfill: resolver },
+				},
 			} );
 
 			const value = registry.select( 'demo' ).getValue( 'arg1', 'arg2' );
@@ -252,32 +244,32 @@ describe( 'createRegistry', () => {
 				return { type: 'SET_PAGE', page, result: [] };
 			} );
 
-			const store = registry.registerReducer( 'demo', ( state = {}, action ) => {
-				switch ( action.type ) {
-					case 'SET_PAGE':
-						return {
-							...state,
-							[ action.page ]: action.result,
-						};
-				}
+			const store = registry.registerStore( 'demo', {
+				reducer: ( state = {}, action ) => {
+					switch ( action.type ) {
+						case 'SET_PAGE':
+							return {
+								...state,
+								[ action.page ]: action.result,
+							};
+					}
 
-				return state;
-			} );
-
-			store.dispatch( { type: 'SET_PAGE', page: 4, result: [] } );
-
-			registry.registerSelectors( 'demo', {
-				getPage: ( state, page ) => state[ page ],
-			} );
-			registry.registerResolvers( 'demo', {
-				getPage: {
-					fulfill,
-					isFulfilled( state, page ) {
-						return state.hasOwnProperty( page );
+					return state;
+				},
+				selectors: {
+					getPage: ( state, page ) => state[ page ],
+				},
+				resolvers: {
+					getPage: {
+						fulfill,
+						isFulfilled( state, page ) {
+							return state.hasOwnProperty( page );
+						},
 					},
 				},
 			} );
 
+			store.dispatch( { type: 'SET_PAGE', page: 4, result: [] } );
 			registry.select( 'demo' ).getPage( 1 );
 			registry.select( 'demo' ).getPage( 2 );
 
@@ -307,14 +299,16 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'should resolve action to dispatch', () => {
-			registry.registerReducer( 'demo', ( state = 'NOTOK', action ) => {
-				return action.type === 'SET_OK' ? 'OK' : state;
-			} );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: () => ( { type: 'SET_OK' } ),
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'NOTOK', action ) => {
+					return action.type === 'SET_OK' ? 'OK' : state;
+				},
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: () => ( { type: 'SET_OK' } ),
+				},
 			} );
 
 			const promise = subscribeUntil( [
@@ -328,14 +322,16 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'should resolve promise action to dispatch', () => {
-			registry.registerReducer( 'demo', ( state = 'NOTOK', action ) => {
-				return action.type === 'SET_OK' ? 'OK' : state;
-			} );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: () => Promise.resolve( { type: 'SET_OK' } ),
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'NOTOK', action ) => {
+					return action.type === 'SET_OK' ? 'OK' : state;
+				},
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: () => Promise.resolve( { type: 'SET_OK' } ),
+				},
 			} );
 
 			const promise = subscribeUntil( [
@@ -350,20 +346,22 @@ describe( 'createRegistry', () => {
 
 		it( 'should resolve promise non-action to dispatch', ( done ) => {
 			let shouldThrow = false;
-			registry.registerReducer( 'demo', ( state = 'OK' ) => {
-				if ( shouldThrow ) {
-					throw 'Should not have dispatched';
-				}
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'OK' ) => {
+					if ( shouldThrow ) {
+						throw 'Should not have dispatched';
+					}
 
-				return state;
+					return state;
+				},
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: () => Promise.resolve(),
+				},
 			} );
 			shouldThrow = true;
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: () => Promise.resolve(),
-			} );
 
 			registry.select( 'demo' ).getValue();
 
@@ -373,14 +371,16 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'should not dispatch resolved promise action on subsequent selector calls', () => {
-			registry.registerReducer( 'demo', ( state = 'NOTOK', action ) => {
-				return action.type === 'SET_OK' && state === 'NOTOK' ? 'OK' : 'NOTOK';
-			} );
-			registry.registerSelectors( 'demo', {
-				getValue: ( state ) => state,
-			} );
-			registry.registerResolvers( 'demo', {
-				getValue: () => Promise.resolve( { type: 'SET_OK' } ),
+			registry.registerStore( 'demo', {
+				reducer: ( state = 'NOTOK', action ) => {
+					return action.type === 'SET_OK' && state === 'NOTOK' ? 'OK' : 'NOTOK';
+				},
+				selectors: {
+					getValue: ( state ) => state,
+				},
+				resolvers: {
+					getValue: () => Promise.resolve( { type: 'SET_OK' } ),
+				},
 			} );
 
 			const promise = subscribeUntil( () => registry.select( 'demo' ).getValue() === 'OK' );
@@ -425,13 +425,14 @@ describe( 'createRegistry', () => {
 
 	describe( 'select', () => {
 		it( 'registers multiple selectors to the public API', () => {
-			const store = registry.registerReducer( 'reducer1', () => 'state1' );
 			const selector1 = jest.fn( () => 'result1' );
 			const selector2 = jest.fn( () => 'result2' );
-
-			registry.registerSelectors( 'reducer1', {
-				selector1,
-				selector2,
+			const store = registry.registerStore( 'reducer1', {
+				reducer: () => 'state1',
+				selectors: {
+					selector1,
+					selector2,
+				},
 			} );
 
 			expect( registry.select( 'reducer1' ).selector1() ).toEqual( 'result1' );
@@ -445,9 +446,11 @@ describe( 'createRegistry', () => {
 	describe( 'subscribe', () => {
 		it( 'registers multiple selectors to the public API', () => {
 			let incrementedValue = null;
-			const store = registry.registerReducer( 'myAwesomeReducer', ( state = 0 ) => state + 1 );
-			registry.registerSelectors( 'myAwesomeReducer', {
-				globalSelector: ( state ) => state,
+			const store = registry.registerStore( 'myAwesomeReducer', {
+				reducer: ( state = 0 ) => state + 1,
+				selectors: {
+					globalSelector: ( state ) => state,
+				},
 			} );
 			const unsubscribe = registry.subscribe( () => {
 				incrementedValue = registry.select( 'myAwesomeReducer' ).globalSelector();
@@ -469,7 +472,9 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'snapshots listeners on change, avoiding a later listener if subscribed during earlier callback', () => {
-			const store = registry.registerReducer( 'myAwesomeReducer', ( state = 0 ) => state + 1 );
+			const store = registry.registerStore( 'myAwesomeReducer', {
+				reducer: ( state = 0 ) => state + 1,
+			} );
 			const secondListener = jest.fn();
 			const firstListener = jest.fn( () => {
 				subscribeWithUnsubscribe( secondListener );
@@ -483,7 +488,9 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'snapshots listeners on change, calling a later listener even if unsubscribed during earlier callback', () => {
-			const store = registry.registerReducer( 'myAwesomeReducer', ( state = 0 ) => state + 1 );
+			const store = registry.registerStore( 'myAwesomeReducer', {
+				reducer: ( state = 0 ) => state + 1,
+			} );
 			const firstListener = jest.fn( () => {
 				secondUnsubscribe();
 			} );
@@ -498,7 +505,9 @@ describe( 'createRegistry', () => {
 		} );
 
 		it( 'does not call listeners if state has not changed', () => {
-			const store = registry.registerReducer( 'unchanging', ( state = {} ) => state );
+			const store = registry.registerStore( 'unchanging', {
+				reducer: ( state = {} ) => state,
+			} );
 			const listener = jest.fn();
 			subscribeWithUnsubscribe( listener );
 
@@ -510,23 +519,22 @@ describe( 'createRegistry', () => {
 
 	describe( 'dispatch', () => {
 		it( 'registers actions to the public API', () => {
-			const store = registry.registerReducer( 'counter', ( state = 0, action ) => {
-				if ( action.type === 'increment' ) {
-					return state + action.count;
-				}
-				return state;
-			} );
 			const increment = ( count = 1 ) => ( { type: 'increment', count } );
-			registry.registerActions( 'counter', {
-				increment,
+			const store = registry.registerStore( 'counter', {
+				reducer: ( state = 0, action ) => {
+					if ( action.type === 'increment' ) {
+						return state + action.count;
+					}
+					return state;
+				},
+				actions: {
+					increment,
+				},
 			} );
 
 			registry.dispatch( 'counter' ).increment(); // state = 1
 			registry.dispatch( 'counter' ).increment( 4 ); // state = 5
 			expect( store.getState() ).toBe( 5 );
-
-			// This uses deprecated functions and will produce a warning.
-			expect( console ).toHaveWarned();
 		} );
 	} );