+++ test/e2e/specs/sidebar.test.js
@@ -2,8 +2,10 @@
  * Internal dependencies
  */
 import {
+	findSidebarPanelWithTitle,
 	newPost,
 	observeFocusLoss,
+	openDocumentSettingsSidebar,
 	pressWithModifier,
 	setViewport,
 } from '../support/utils';
@@ -12,12 +14,12 @@ const SIDEBAR_SELECTOR = '.edit-post-sidebar';
 const ACTIVE_SIDEBAR_TAB_SELECTOR = '.edit-post-sidebar__panel-tab.is-active';
 const ACTIVE_SIDEBAR_BUTTON_TEXT = 'Document';
 
-describe( 'Publishing', () => {
+describe( 'Sidebar', () => {
 	beforeAll( () => {
 		observeFocusLoss();
 	} );
 
-	it( 'Should have sidebar visible at the start with document sidebar active on desktop', async () => {
+	it( 'should have sidebar visible at the start with document sidebar active on desktop', async () => {
 		await setViewport( 'large' );
 		await newPost();
 		const { nodesCount, content, height, width } = await page.$$eval( ACTIVE_SIDEBAR_TAB_SELECTOR, ( nodes ) => {
@@ -41,14 +43,14 @@ describe( 'Publishing', () => {
 		expect( height ).toBeGreaterThan( 10 );
 	} );
 
-	it( 'Should have the sidebar closed by default on mobile', async () => {
+	it( 'should have the sidebar closed by default on mobile', async () => {
 		await setViewport( 'small' );
 		await newPost();
 		const sidebar = await page.$( SIDEBAR_SELECTOR );
 		expect( sidebar ).toBeNull();
 	} );
 
-	it( 'Should close the sidebar when resizing from desktop to mobile', async () => {
+	it( 'should close the sidebar when resizing from desktop to mobile', async () => {
 		await setViewport( 'large' );
 		await newPost();
 
@@ -62,7 +64,7 @@ describe( 'Publishing', () => {
 		expect( sidebarsMobile ).toHaveLength( 0 );
 	} );
 
-	it( 'Should reopen sidebar the sidebar when resizing from mobile to desktop if the sidebar was closed automatically', async () => {
+	it( 'should reopen sidebar the sidebar when resizing from mobile to desktop if the sidebar was closed automatically', async () => {
 		await setViewport( 'large' );
 		await newPost();
 		await setViewport( 'small' );
@@ -76,14 +78,14 @@ describe( 'Publishing', () => {
 		expect( sidebarsDesktop ).toHaveLength( 1 );
 	} );
 
-	it( 'Should preserve tab order while changing active tab', async () => {
+	it( 'should preserve tab order while changing active tab', async () => {
 		await newPost();
 
 		// Region navigate to Sidebar.
-		await pressWithModifier( 'Control', '`' );
-		await pressWithModifier( 'Control', '`' );
-		await pressWithModifier( 'Control', '`' );
-		await pressWithModifier( 'Control', '`' );
+		await pressWithModifier( 'ctrl', '`' );
+		await pressWithModifier( 'ctrl', '`' );
+		await pressWithModifier( 'ctrl', '`' );
+		await pressWithModifier( 'ctrl', '`' );
 
 		// Tab lands at first (presumed selected) option "Document".
 		await page.keyboard.press( 'Tab' );
@@ -101,5 +103,33 @@ describe( 'Publishing', () => {
 			document.activeElement.classList.contains( 'is-active' )
 		) );
 		expect( isActiveBlockTab ).toBe( true );
+	} );
+
+	it( 'should be possible to programmatically remove Document Settings panels', async () => {
+		await newPost();
+
+		await openDocumentSettingsSidebar();
+
+		expect( await findSidebarPanelWithTitle( 'Categories' ) ).toBeDefined();
+		expect( await findSidebarPanelWithTitle( 'Tags' ) ).toBeDefined();
+		expect( await findSidebarPanelWithTitle( 'Featured Image' ) ).toBeDefined();
+		expect( await findSidebarPanelWithTitle( 'Excerpt' ) ).toBeDefined();
+		expect( await findSidebarPanelWithTitle( 'Discussion' ) ).toBeDefined();
+
+		await page.evaluate( () => {
+			const { removeEditorPanel } = wp.data.dispatch( 'core/edit-post' );
+
+			removeEditorPanel( 'taxonomy-panel-category' );
+			removeEditorPanel( 'taxonomy-panel-post_tag' );
+			removeEditorPanel( 'featured-image' );
+			removeEditorPanel( 'post-excerpt' );
+			removeEditorPanel( 'discussion-panel' );
+		} );
+
+		expect( await findSidebarPanelWithTitle( 'Categories' ) ).toBeUndefined();
+		expect( await findSidebarPanelWithTitle( 'Tags' ) ).toBeUndefined();
+		expect( await findSidebarPanelWithTitle( 'Featured Image' ) ).toBeUndefined();
+		expect( await findSidebarPanelWithTitle( 'Excerpt' ) ).toBeUndefined();
+		expect( await findSidebarPanelWithTitle( 'Discussion' ) ).toBeUndefined();
 	} );
 } );